Add date of London upgrade to Gas and Fees page
Now that the London upgrade is live, we should make it clear that the `Prior to the London Upgrade` section is now describing the *past*, and the `After the London Upgrade` is now describing the *current* Gas and Fees method. I suggest making this clearer by adding the date that the london upgrade went live. Without the date, if you didn't know what the London upgrade was or if/when it would go live, you'd be unsure which of those Gas and Fees methods is the *current* method, without researching more about the London upgrade.

Simple minor change, feel free to modify/reject if not ideal.